% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor.R
\name{mixcuref_fit}
\alias{mixcuref_fit}
\title{Adapter for mixcure.meta: fit from a cure_spec}
\usage{
mixcuref_fit(spec, control = list())
}
\arguments{
\item{spec}{A mc_spec object}

\item{control}{A list of engine-specific controls}
}
\value{
An engine-specific fitted object
}
\description{
Adapter for mixcure.meta: fit from a cure_spec
}
